# [Epic-1] Story 1.5: Performance Validation