## Summary
Implemented comprehensive performance validation benchmarks for Epic 1 foundation.

## Epic/Story Reference
- Epic: 1 - Foundation & Core Data Pipeline
- Story: 1.5 - Basic Performance Validation

## Performance Results
All targets exceeded by significant margins:
- **Tick insertion**: 588K/s (target: 10K/s) - 58x better
- **Memory for 1M ticks**: 67MB (target: <500MB) - 7.5x under
- **Query response**: 1.3ms (target: <100ms) - 77x faster
- **CSV import**: 476K/s (target: 10K/s) - 47x better

## Changes
- Created 4 benchmark suites using Criterion.rs
- Added performance baseline report
- Validated all Epic 1 performance targets

## Testing
```bash
cargo bench                    # Run all benchmarks
cargo test --release           # Run all tests
cargo clippy --all-targets     # Lint checks pass
```

## Checklist
- [x] All acceptance criteria met
- [x] Tests passing
- [x] Documentation updated
- [x] Performance targets validated
- [x] Code quality checks pass

Implements Story 1.5